Seperated code into different JS files for readability
