+++ index.html
@@ -25,6 +25,6 @@
   </div>
 
 
-  <script src="app.js"></script>
+  <script src="JS/app.js" type="module"></script>
 </body>
 </html>
